Fix fix_overlap
The overlapping path must be reset (1, 1) before the split_1 path is
unset (0, -1), in case the share some connections. This shouldn't be the
case, but the overlap_path building is not totally correct.

The overlap_path can really be two separate parts of the same path,
which "creates" non-existing connections. This should be fixed, even if
this solution works (it is fucking ugly).